refactor(cli): split hook into api/hook leaf shape
Reorganize the flat `api/hook/hook.mjs` into the fractal leaf shape mirroring
`api/theme/build`:

- hook.mjs            dispatcher + barrel (routing only; same `hook` export)
- _adapter.mjs        shared resolveCoreDir + resolveHookDoc (list + detail)
- list/list.mjs       -> hook.list (data.detail names|compact|full)
- detail/detail.mjs   -> hook.detail
- detail/params/params.mjs -> hook.detail.params

Pure reorg: `--json` and human output are byte-identical across every mode
(list names/compact/full, single-hook detail, --params, and the unknown-hook
error, verified stdout+stderr+exit). api/index.mjs and the CLI command handler
are unchanged. Types stay central in types/hook.d.ts (now re-exporting HookDoc/
HookParamDoc for precise leaf @returns). Adds colocated detail/params leaf tests
(the list leaf is already covered by detail-levels.test.mjs).
